:clown_face: Refactor StUF-BG mocks in tests
